server2/requirements.txt
fastapi
uvicorn
pydantic
pymongo
kafka-python